separated the classes in model and view
