+++ src/cm-chessboard/model/Position.js
