+++ src/cm-chessboard/view/PositionAnimationsQueue.js
@@ -3,8 +3,8 @@
  * Repository: https://github.com/shaack/cm-chessboard
  * License: MIT, see file 'LICENSE'
  */
-import {FEN_EMPTY_POSITION, Position} from "./Position.js"
-import {Svg} from "./View.js"
+import {FEN_EMPTY_POSITION, Position} from "../model/Position.js"
+import {Svg} from "./ChessboardView.js"
 
 /*
 * Thanks to markosyan for the idea to the PromiseQueue
